Adjust structure of the repository and add Landing documentation and user documentation for MMR Mobile
