--- Applications/MmrMobile/Samples/C#/Tests/appsettings.json
@@ -1,7 +1,0 @@
-{
-  "HomagApiGateway": {
-    "BaseUrl": "https://connect.homag.cloud",
-    "Username": "820a547d-b2b0-4e0e-9b0f-2533e8310b3e", // Keep this username for your requests
-    "Token": "" // Use your personal access token from tapio"
-  }
-}
