+++ Applications/MmrMobile/Samples/CSharp/Console/HomagConnect.MmrMobile.Samples.Console.csproj
