+++ Applications/MmrMobile/Samples/CSharp/Console/StateGroupCodes.cs
