+++ Applications/MmrMobile/Samples/CSharp/Tests/HomagConnect.MmrMobile.Samples.Tests.csproj
