+++ Applications/MmrMobile/Samples/CSharp/Tests/appsettings.json
@@ -1,0 +1,9 @@
+{
+  "HomagConnect": {
+    "BaseUrl": "https://connect.homag.cloud",
+    "MmrMobile": {
+      "SubscriptionId": "", // Keep this username for your requests
+      "Token": "" // Use your personal access token from tapio"
+    }
+  }
+}
